requirements.txt
--extra-index-url https://www.paddlepaddle.org.cn/packages/stable/cpu/

paddlepaddle==3.2.0
paddleocr[all]

llama-cpp-python
pdfplumber
python-docx
docx2pdf
tqdm
Pillow
numpy
opencv-python
scikit-image
openai-whisper
speechbrain
librosa
scikit-learn
torch
python-magic
python-magic-bin
soundfile
torchaudio